1) Clears all physical memory 2) Also fixes bug in avoiding remapping of pte_t 3) Merged new changes of Makefile

+++ include/stdio.h
@@ -1,5 +1,6 @@
 #ifndef _STDIO_H
 #define _STDIO_H
+#include <defs.h>
 
 char *itoa(uint64_t val, char *str, int32_t base);
 void putchar(char mychar);
@@ -8,8 +9,8 @@ int32_t kprintf(const char *str, ...);
 
 int32_t scanf(const char *format, ...);
 
-void *memcpy(void *destination, void *source, int num);
-void *memset(void *ptr, int value, int num);
+void *memcpy(void *destination, void *source, uint64_t num);
+void *memset(void *ptr, int value, uint64_t num);
 
 int32_t pow(int base, int power);
 int32_t atoi(char *p);
